+++ src/App.jsx
@@ -1,10 +1,7 @@
 import './App.scss'
 
 function App() {
-  return (
-    <div className="App">
-    </div>
-  )
+  return <div className="App"></div>
 }
 
 export default App
