+++ src/UIComponents/admin/UserCardItem.module.scss
@@ -1,5 +1,5 @@
-@use '../styles/mixin';
-@use '../styles/var' as color;
+@use '../...styles/mixin';
+@use '../../styles/var' as color;
 
 .card__container {
   width: 15.3125rem;
