+++ src/UIComponents/buttons/ButtonUI.module.scss
@@ -1,5 +1,5 @@
-@use '../styles/var' as var;
-@use '../styles/mixin';
+@use '../../styles/var' as var;
+@use '../../styles/mixin';
 
 .btn__pill__large {
   width: 100%;
