+++ src/UIComponents/inputs/Input.module.scss
@@ -1,5 +1,5 @@
-@use '../styles/var' as color;
-@use '../styles/mixin';
+@use '../../styles/var' as color;
+@use '../../styles/mixin';
 
 .input__container {
   position: relative;
