+++ src/UIComponents/listItems/FollowListItem.module.scss
@@ -1,5 +1,5 @@
-@use '../styles/var' as color;
-@use '../styles/mixin';
+@use '../../styles/var' as color;
+@use '../../styles/mixin';
 
 .listItem__container {
   width: 100%;
